+++ src/index.d.ts
@@ -44,6 +44,24 @@ import {
 } from './services/config.js';
 
 import {
+	fetchArtistBySlug,
+	fetchArtistLessons,
+	fetchArtists
+} from './services/content/artist.ts';
+
+import {
+	fetchGenreBySlug,
+	fetchGenreLessons,
+	fetchGenres
+} from './services/content/genre.ts';
+
+import {
+	fetchInstructorBySlug,
+	fetchInstructorLessons,
+	fetchInstructors
+} from './services/content/instructor.ts';
+
+import {
 	enrollUserInGuidedCourse,
 	fetchEnrollmentPageMetadata,
 	guidedCourses,
@@ -61,7 +79,6 @@ import {
 	mapContentToParent,
 	resetAllLearningPaths,
 	startLearningPath,
-	updateActivePath,
 	updateDailySession
 } from './services/content-org/learning-paths.ts';
 
@@ -88,6 +105,7 @@ import {
 	getLegacyMethods,
 	getLessonContentRows,
 	getNewAndUpcoming,
+	getOwnedContent,
 	getRecent,
 	getRecommendedForYou,
 	getScheduleContentRows,
@@ -115,7 +133,6 @@ import {
 	getAllStartedOrCompleted,
 	getLastInteractedOf,
 	getNavigateTo,
-	getNextLesson,
 	getProgressDataByIds,
 	getProgressState,
 	getProgressStateByIds,
@@ -141,6 +158,7 @@ import {
 
 import {
 	createForumCategory,
+	deleteForumCategory,
 	fetchForumCategories,
 	updateForumCategory
 } from './services/forums/categories.ts';
@@ -253,21 +271,23 @@ import {
 } from './services/recommendations.js';
 
 import {
+	getReportIssueOptions,
+	report
+} from './services/reporting/reporting.ts';
+
+import {
+	buildEntityAndTotalQuery,
 	fetchAll,
 	fetchAllFilterOptions,
 	fetchAllPacks,
-	fetchArtistLessons,
-	fetchArtists,
 	fetchByRailContentId,
 	fetchByRailContentIds,
 	fetchByReference,
 	fetchChatAndLiveEnvent,
-	fetchCoachLessons,
 	fetchComingSoon,
 	fetchCommentModContentData,
 	fetchContentRows,
 	fetchFoundation,
-	fetchGenreLessons,
 	fetchHierarchy,
 	fetchLeaving,
 	fetchLessonContent,
@@ -280,9 +300,11 @@ import {
 	fetchMethodPreviousNextLesson,
 	fetchMethodV2IntroVideo,
 	fetchMethodV2Structure,
+	fetchMethodV2StructureFromId,
 	fetchNewReleases,
 	fetchNextPreviousLesson,
 	fetchOtherSongVersions,
+	fetchOwnedContent,
 	fetchPackAll,
 	fetchPackData,
 	fetchPlayAlongsCount,
@@ -302,6 +324,7 @@ import {
 	fetchTabData,
 	fetchTopLevelParentId,
 	fetchUpcomingEvents,
+	getSanityDate,
 	getSortOrder,
 	jumpToContinueContent
 } from './services/sanity.js';
@@ -434,6 +457,7 @@ declare module 'musora-content-services' {
 		assignModeratorToComment,
 		blockUser,
 		blockedUsers,
+		buildEntityAndTotalQuery,
 		buildCertificateData,
 		buildImageSRC,
 		calculateLongestStreaks,
@@ -454,6 +478,7 @@ declare module 'musora-content-services' {
 		createThread,
 		deleteAccount,
 		deleteComment,
+		deleteForumCategory,
 		deleteItemsFromPlaylist,
 		deleteNotification,
 		deletePicture,
@@ -472,6 +497,7 @@ declare module 'musora-content-services' {
 		fetchAllCompletedStates,
 		fetchAllFilterOptions,
 		fetchAllPacks,
+		fetchArtistBySlug,
 		fetchArtistLessons,
 		fetchArtists,
 		fetchAwardsForUser,
@@ -482,7 +508,6 @@ declare module 'musora-content-services' {
 		fetchCertificate,
 		fetchChatAndLiveEnvent,
 		fetchChatSettings,
-		fetchCoachLessons,
 		fetchComingSoon,
 		fetchComment,
 		fetchCommentModContentData,
@@ -499,9 +524,14 @@ declare module 'musora-content-services' {
 		fetchFollowedThreads,
 		fetchForumCategories,
 		fetchFoundation,
+		fetchGenreBySlug,
 		fetchGenreLessons,
+		fetchGenres,
 		fetchHandler,
 		fetchHierarchy,
+		fetchInstructorBySlug,
+		fetchInstructorLessons,
+		fetchInstructors,
 		fetchInterests,
 		fetchLastInteractedChild,
 		fetchLatestThreads,
@@ -520,12 +550,14 @@ declare module 'musora-content-services' {
 		fetchMethodPreviousNextLesson,
 		fetchMethodV2IntroVideo,
 		fetchMethodV2Structure,
+		fetchMethodV2StructureFromId,
 		fetchNewReleases,
 		fetchNextContentDataForParent,
 		fetchNextPreviousLesson,
 		fetchNotificationSettings,
 		fetchNotifications,
 		fetchOtherSongVersions,
+		fetchOwnedContent,
 		fetchPackAll,
 		fetchPackData,
 		fetchPlayAlongsCount,
@@ -596,6 +628,7 @@ declare module 'musora-content-services' {
 		getNewlyEarnedAwards,
 		getNextLesson,
 		getOnboardingRecommendedContent,
+		getOwnedContent,
 		getPracticeNotes,
 		getPracticeSessions,
 		getProgressDataByIds,
@@ -605,7 +638,9 @@ declare module 'musora-content-services' {
 		getRecent,
 		getRecentActivity,
 		getRecommendedForYou,
+		getReportIssueOptions,
 		getResumeTimeSecondsByIds,
+		getSanityDate,
 		getScheduleContentRows,
 		getSortOrder,
 		getStartedOrCompletedProgressOnly,
@@ -664,6 +699,7 @@ declare module 'musora-content-services' {
 		removeContentAsNotInterested,
 		removeUserPractice,
 		replyToComment,
+		report,
 		reportComment,
 		reportPlaylist,
 		requestEmailChange,