+++ test/initializeTests.js
@@ -51,7 +51,9 @@ export async function initializeTestService(useLive = false, isAdmin = false) {
     isMA: true,
   }
   initializeService(config)
-  let mock = jest.spyOn(railContentModule, 'fetchUserPermissionsData')
-  let testData = { permissions: [78, 91, 92], isAdmin: isAdmin }
-  mock.mockImplementation(() => testData)
+
+  // Mock user permissions
+  let permissionsMock = jest.spyOn(railContentModule, 'fetchUserPermissionsData')
+  let permissionsData = { permissions: [108, 91, 92], isAdmin: isAdmin }
+  permissionsMock.mockImplementation(() => permissionsData)
 }